XARATINGS-1 Implement the ratings application initial feature set
Force the "use" field of JSX and CSS extensions to be the correct ones when building the final XAR, regardless
of the values they have in the sources.
This is to avoid programmers mistakes, as it is very easy to forget to check they have the proper ones 
when commiting changes based on a XAR export.



git-svn-id: https://svn.xwiki.org/svnroot/xwiki/sandbox/xwiki-application-ratings@14393 f329d543-caf0-0310-9063-dda96c69346f