XARATINGS-1 Implement the ratings application initial feature set
[refactoring]
I removed the notion of container from the ratings plugin, to get rid of the dependency on the xwiki-comments plugin. The container interface brings currently no value in the ratings plugin, since in the end only documents can be rated. Containers were complexifying the code for no good reason, since the container interface from the comments plugin is planned to be refactored anyway (for example, using the work that will be done on the annotation feature, see http://dev.xwiki.org/xwiki/bin/view/Design/AnnotationFeature)
Getting rid of that dependency allows us to release the ratings plugin individually from the comments plugin (which itself will need some adaptation to work with current version of XWiki core: it is written for the core-1.5-curriki5 branch, itself patched), but leaving us the possibility of having ratings and comments plugin work together, for example to rate comments in the context of comments stored as separate pages (i.e. one object per comment, not in the commented document itself).
Note that this refactoring keeps the exact same functionnalities the ratings plugin already had.


git-svn-id: https://svn.xwiki.org/svnroot/xwiki/sandbox/plugins/ratings/trunk@14421 f329d543-caf0-0310-9063-dda96c69346f